some refactoring, style updates and putting the title and logos into the conf file
