+++ static/sass/_variables.scss
@@ -5,6 +5,6 @@ $verydarkblue: #275F90;
 $darkblue: #1070B7;
 $blue: #74C5EB;
 $lightblue: #DAF0F9;
-$red: #f0586a;
+$red: #ff7979;
 $green: #7ed321;
 $background: #F5F9FA;
