--- templates/admin.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Cineca Services Registry</title>
-        <meta charset="utf-8" />
-        <meta name="viewport" content="initial-scale=1.0, width=device-width, shrink-to-fit=no"/>
-
-        <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.10/css/all.css"
-              integrity="sha384-+d0P83n9kaQMCwj8F4RJB66tzIwOKmrdb46+porD/OvrJ+37WqIM7UoBtwHO6Nlg"
-              crossorigin="anonymous" />
-
-        <link rel="stylesheet"
-	      href="https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,300;0,400;0,600;0,700;0,800;1,300;1,400;1,600;1,700;1,800&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" />
-
-        <link rel="stylesheet" href="{{ static('css/style.css') }}" />
-    </head>
-    <body>
-
-      <nav>
-        <a href="{{ url('admin') }}">
-  	       ADMIN
-        </a>
-      </nav>
-
-      <main>
-        {% if message %}
-          <p class="message {{ message_type }}">
-            {{ message }}
-          </p>
-        {% endif %}
-
-        <form action="{{ url('admin') }}" method="POST">
-          <div>
-            <input type="url"
-            		   name="new_url"
-            		   placeholder="Input the URL that you want add" />
-            <button type="submit"><i class="fas fa-plus"></i></button>
-          </div>
-
-          <div id="cards-container" data-cards="{{ services|length }}">
-            <div class="card">
-              <section class="right-part">
-                <h1>New Beacon Instance</h1>
-                <input type="text" name="new_url" placeholder="Enter the Beacon Instance URL" />
-                
-                <input type="submit" value="ADD" />
-              </section>
-            </div>
-            {% for service in services %}
-            {% if not service.error %}
-              {% with remove_enabled = True %}
-                {% include "card.html" %}
-              {% endwith %}
-            {% endif %}
-              {# Couldn't contact: {{ service }} #}
-            {% endfor %}
-          </div>
-
-        </form>
-
-      </main>
-
-      <footer>
-        <section id="EGA">
-          <p>Hosted by the <a href="https://ega-archive.org/">EGA</a></p>
-        </section>
-        <section id="EU">
-          <p>This project has received funding from the European Union's Horizon 2020 Research and Innovation Programme under grant agreement No. 825775.</p>
-          <img src="{{ static('img/eu_white.jpg') }}" alt="EU Logo" />
-        </section>
-        <!-- <p>&copy; Copyright 2020</p> -->
-      </footer>
-
-    </body>
-</html>
